Refactor shifted quantization
* refactoring shifted quantization function

* small refactoring and renaming

* Code cleaning and reorganising in files
