+++ src/operator/quantization/quantize_graph_pass.cc
@@ -19,20 +19,11 @@
 
 /*!
  *  Copyright (c) 2016 by Contributors
- * \file quantization.cc
+ * \file quantize_graph_pass.cc
  * \brief
  */
 
-#include <mxnet/op_attr_types.h>
-#include <nnvm/graph.h>
-#include <nnvm/pass.h>
-#include <queue>
-#include <unordered_map>
-#include <unordered_set>
-#include <vector>
-#include "quantize_v2-inl.h"
-#include "../nn/mkldnn/mkldnn_fully_connected-inl.h"
-#include "../../common/utils.h"
+#include "quantize_graph_pass.h"
 
 namespace mxnet {
 namespace op {
@@ -54,20 +45,6 @@ static inline size_t GetNumOutputs(ObjectPtr node) {
     num_outputs = num_visible_output_func(node->attrs);
   }
   return num_outputs;
-}
-
-ObjectPtr CreateNode(std::string op_name, std::string node_name) {
-  ObjectPtr node = Node::Create();
-  node->attrs.name = node_name;
-  if (op_name == "nullptr") {
-    node->attrs.op = nullptr;
-    // ugly workaround because VariableParam is not exposed
-    node->attrs.parsed =
-      nnvm::Symbol::CreateVariable(node->attrs.name).outputs[0].node->attrs.parsed;
-  } else {
-    node->attrs.op = Op::Get(op_name);
-  }
-  return node;
 }
 
 /*!
@@ -579,204 +556,6 @@ Graph SetCalibTableToQuantizedGraph(Graph&& g) {
   return g;
 }
 
-static NDArray* FindInArgByName(const Graph &g, const std::string& name) {
-  const std::vector<std::string>& in_arg_names =
-      g.GetAttr<std::vector<std::string>>("in_arg_names");
-  size_t i = std::distance(in_arg_names.begin(),
-                           std::find(in_arg_names.begin(), in_arg_names.end(), name));
-  if (i == in_arg_names.size()) {
-    LOG(FATAL) << name << " not found in in_arg_names";
-  }
-  return g.GetAttr<NDArray **>("in_args")[i];
-}
-
-template <bool require_bias>
-static inline bool IsOneDNNFullyConnected(const ObjectPtr& n) {
-#if MXNET_USE_MKLDNN == 1
-  if (n->op() == Op::Get("_sg_mkldnn_fully_connected")) {
-    auto const& param = nnvm::get<MKLDNNFCFullParam>(n->attrs.parsed);
-    if (!(param.mkldnn_param.channel_wise_quantize.has_value() &&
-          param.mkldnn_param.channel_wise_quantize.value())) {
-      return !require_bias || (param.default_param.no_bias == false &&
-                               n->inputs[2].node->is_variable());
-    }
-  }
-#endif
-  return false;
-}
-
-static inline bool IsQuantize(const ObjectPtr& n) {
-  if (n->op() == Op::Get("_contrib_quantize_v2")) {
-    auto const &param = nnvm::get<QuantizeV2Param>(n->attrs.parsed);
-    if (param.min_calib_range.has_value() &&
-        param.min_calib_range.value() < 0.0f) {
-      return true;
-    }
-  }
-  return false;
-}
-
-// Rescales weights, min_weight and max_weight. Returns bias_int32_rescale.
-static inline float RescaleWeights(const Graph &g, const ObjectPtr &fc, NDArray* weight_tensor) {
-  ObjectPtr &quantize = fc->inputs[0].node;
-  auto min_data = std::stof(quantize->attrs.dict.at("min_calib_range"));
-  auto max_data = std::stof(quantize->attrs.dict.at("max_calib_range"));
-
-  float *min_weight = FindInArgByName(g, fc->inputs[5].node->attrs.name)->data().dptr<float>();
-  float *max_weight = FindInArgByName(g, fc->inputs[6].node->attrs.name)->data().dptr<float>();
-  float min_bias = *FindInArgByName(g, fc->inputs[7].node->attrs.name)->data().dptr<float>();
-  float max_bias = *FindInArgByName(g, fc->inputs[8].node->attrs.name)->data().dptr<float>();
-
-  float data_scale_ = kUint8Range / (max_data - min_data);
-  float weight_scale = GetQuantizeScale(mshadow::kInt8, *min_weight, *max_weight);
-  float bias_scale = GetQuantizeScale(mshadow::kInt8, min_bias, max_bias);
-  float bias_int32_rescale = data_scale_ * weight_scale / bias_scale;
-
-  // // TODO(zhennan): mkldnn has bug to handle INT_MAX in bias, so set the
-  // // maximum value of bias to INT_MAX / 2.
-  float bias_max_rescale = mshadow::red::limits::MaxValue<int32_t>() / 2 /
-                           MaxAbs(min_bias, max_bias) / bias_scale;
-  if (bias_int32_rescale > bias_max_rescale) {
-    LOG(INFO) << "RESCALING WEIGHTS in shifted quantization because bias scale "
-                 "is too big in layer " << fc->attrs.name;
-    // avoid overflow on bias
-    bias_int32_rescale = bias_max_rescale;
-    float weight_rescale =
-        bias_int32_rescale * bias_scale / data_scale_ / weight_scale;
-
-    size_t weight_size = weight_tensor->shape().Size();
-    int8_t *weight_ptr = weight_tensor->data().dptr<int8_t>();
-    for (int32_t i = 0; i < static_cast<int32_t>(weight_size); ++i) {
-      weight_ptr[i] = std::round(weight_ptr[i] * weight_rescale);
-    }
-    *min_weight *= weight_rescale;
-    *max_weight *= weight_rescale;
-  }
-  return bias_int32_rescale;
-}
-
-static inline void ShiftBias(int32_t* bias_ptr_int32, size_t bias_size,
-                             NDArray* weight_tensor, int32_t shift_value) {
-  CHECK_EQ(static_cast<size_t>(weight_tensor->shape()[0]), bias_size);
-  int8_t* weight_ptr = weight_tensor->data().dptr<int8_t>();
-  for (dim_t i = 0; i < weight_tensor->shape()[0]; ++i) {
-    for (dim_t j = 0; j < weight_tensor->shape()[1]; j++) {
-      bias_ptr_int32[i] -= shift_value * (*weight_ptr++);
-    }
-  }
-}
-
-Graph OneDNNShiftedQuantization(Graph&& g) {
-  bool disable_shifted_quant =
-      dmlc::GetEnv("MXNET_DISABLE_SHIFTED_QUANTIZATION_OPTIMIZATIONS", true);
-  bool quantize_fc = !dmlc::GetEnv("MXNET_DISABLE_SHIFTED_QUANTIZE_FC_OPTIMIZATION", false);
-  bool fc_fc = !dmlc::GetEnv("MXNET_DISABLE_SHIFTED_FC_FC_OPTIMIZATION", false);
-  LOG(INFO) << "Running OneDNN shifted quantization: " << !disable_shifted_quant;
-  // No change to aux params
-  g.attrs["new_aux_names"] = std::make_shared<nnvm::any>(std::vector<std::string>());
-  g.attrs["new_aux"] = std::make_shared<nnvm::any>(std::vector<NDArray *>());
-
-  // New args to replace the old
-  std::vector<std::string> new_arg_names;
-  std::vector<NDArray *> new_arg_vector;
-
-#if MXNET_USE_MKLDNN == 1
-  if (!disable_shifted_quant) {
-    DFSVisit(g.outputs, [&](const ObjectPtr &fc) {
-      // Find Quantize->FC pattern and rescale bias from int8 to int32 and shift
-      if (IsOneDNNFullyConnected<true>(fc)) {
-        auto input0 = fc->inputs[0].node;
-        if (quantize_fc && IsQuantize(input0)) {
-          ObjectPtr &quantize = input0;
-          ObjectPtr& bias_node = fc->inputs[2].node;
-          std::string bias_name_old = bias_node->attrs.name;
-          NDArray* bias_in_arg_ptr = FindInArgByName(g, bias_name_old);
-          if (bias_in_arg_ptr->dtype() != mshadow::kInt8) return;
-          std::string bias_name_s32 = bias_node->attrs.name + "_s32";
-          bias_node = CreateNode("nullptr", bias_name_s32);
-          new_arg_names.push_back(bias_name_s32);
-
-          quantize->attrs.dict["shifted"] = "True";
-          if (quantize->op()->attr_parser) quantize->op()->attr_parser(&(quantize->attrs));
-
-          NDArray *weight_tensor = FindInArgByName(g, fc->inputs[1].node->attrs.name);
-
-          float bias_int32_rescale = RescaleWeights(g, fc, weight_tensor);
-
-          new_arg_vector.push_back(
-              new NDArray(kDefaultStorage, bias_in_arg_ptr->shape(),
-                          Context::CPU(), false, mshadow::kInt32));
-          int32_t *bias_ptr_int32 = new_arg_vector.back()->data().dptr<int32_t>();
-          size_t bias_size = bias_in_arg_ptr->shape().Size();
-          int8_t *bias_ptr_old = bias_in_arg_ptr->data().dptr<int8_t>();
-
-          for (size_t i = 0; i < bias_size; ++i) {
-            bias_ptr_int32[i] = static_cast<int32_t>(
-                std::round(bias_ptr_old[i] * bias_int32_rescale));
-          }
-          float min_data = std::stof(quantize->attrs.dict.at("min_calib_range"));
-          float max_data = std::stof(quantize->attrs.dict.at("max_calib_range"));
-          float data_scale = kUint8Range / (max_data - min_data);
-          int32_t shift_value = static_cast<int32_t>(std::round(data_scale * -min_data));
-          ShiftBias(bias_ptr_int32, bias_size, weight_tensor, shift_value);
-          LOG(INFO) << "applied shifted quantization on QUANTIZE->FC";
-        } else if (fc_fc && IsOneDNNFullyConnected<false>(input0)) {
-          ObjectPtr& first_fc = input0;
-          auto const& param =
-              nnvm::get<MKLDNNFCFullParam>(first_fc->attrs.parsed);
-          // TODO(sfraczek): remove !with_eltwise when onednn version upgrades
-          if (!param.mkldnn_param.with_eltwise) {
-            ObjectPtr& bias_node = fc->inputs[2].node;
-            std::string bias_name_old = bias_node->attrs.name;
-            NDArray* bias_in_arg_ptr = FindInArgByName(g, bias_name_old);
-            if (bias_in_arg_ptr->dtype() != mshadow::kInt8) return;
-            std::string bias_name_s32 = bias_node->attrs.name + "_s32";
-            bias_node = CreateNode("nullptr", bias_name_s32);
-            new_arg_names.push_back(bias_name_s32);
-
-            first_fc->attrs.dict["shifted_output"] = "True";
-            if (first_fc->op()->attr_parser)
-              first_fc->op()->attr_parser(&(first_fc->attrs));
-
-            NDArray* weight_tensor =
-                FindInArgByName(g, fc->inputs[1].node->attrs.name);
-
-            float bias_int32_rescale = RescaleWeights(g, fc, weight_tensor);
-
-            new_arg_vector.push_back(
-                new NDArray(kDefaultStorage, bias_in_arg_ptr->shape(),
-                            Context::CPU(), false, mshadow::kInt32));
-
-            int32_t* bias_ptr_int32 =
-                new_arg_vector.back()->data().dptr<int32_t>();
-            size_t bias_size = bias_in_arg_ptr->shape().Size();
-            int8_t* bias_ptr_old = bias_in_arg_ptr->data().dptr<int8_t>();
-
-            for (size_t i = 0; i < bias_size; ++i) {
-              bias_ptr_int32[i] = static_cast<int32_t>(
-                  std::round(bias_ptr_old[i] * bias_int32_rescale));
-            }
-
-            float min_data =
-                std::stof(first_fc->attrs.dict.at("min_calib_range"));
-            float max_data =
-                std::stof(first_fc->attrs.dict.at("max_calib_range"));
-            float data_scale = kUint8Range / (max_data - min_data);
-            int32_t shift_value =
-                static_cast<int32_t>(std::round(data_scale * -min_data));
-            ShiftBias(bias_ptr_int32, bias_size, weight_tensor, shift_value);
-            LOG(INFO) << "applied shifted quantization on FC->FC";
-          }
-        }
-      }
-    });
-  }
-#endif
-  g.attrs["new_arg_names"] = std::make_shared<nnvm::any>(new_arg_names);
-  g.attrs["new_args"] = std::make_shared<nnvm::any>(new_arg_vector);
-  return g;
-}
-
 NNVM_REGISTER_PASS(QuantizeGraph)
 .describe("")
 .set_body(QuantizeGraph)
@@ -786,11 +565,6 @@ NNVM_REGISTER_PASS(QuantizeGraph)
 NNVM_REGISTER_PASS(SetCalibTableToQuantizedGraph)
 .describe("")
 .set_body(SetCalibTableToQuantizedGraph)
-.set_change_graph(true);
-
-NNVM_REGISTER_PASS(OneDNNShiftedQuantization)
-.describe("Enables shifted quantization.")
-.set_body(OneDNNShiftedQuantization)
 .set_change_graph(true);
 
 }  // namespace op
